pyproject.toml
[project]
name = "flaskr"
version = "0.0.1"
description = "An app made for Orbital"
requires-python = ">=3.8"
dependencies = [
    'Flask>=3.0.3',
    'Flask-Session>=0.8.0',
    'Flask-SQLAlchemy>=3.1.1',
    'psycopg>=3.1.19',
    'pycryptodome>=3.20.0',
    'waitress>=3.0.0',
    'Werkzeug>=3.0.3'
]

[tool.setuptools]
py-modules = []